+++ TODO.md
@@ -48,6 +48,16 @@
 - [ ] 更换[KindEditor](https://www.oschina.net/news/104631/kindeditor-upload-vulnerability)
 
 
+---------------
+
+文档上传流程优化：
+    1. 未登录用户，不允许上传文档
+    2. 已登录用户，积分达不到管理后台规定的积分数，不允许上传文档
+    3. 用户上传文档的时候，前端计算文档 MD5 的值，然后请求后台对比值是否存在，存在了，则表示文档已经存在，直接`秒传`
+    4. 用户上传的文档在管理后台不存在，则把原文档上传到云存储，文档信息(分类、大小、md5等)、用户积分变化等存入数据库
+    5. 数据入库成功之后，再进行文档转换，文档最终转换成功之后，更新文档状态
+
+
 
 
 ## DocHub v2.1
