+++ views/Home/default/Upload/index.html
@@ -56,6 +56,7 @@
                                     <option value="8">8个金币</option>
                                     <option value="10">10个金币</option>
                                     <option value="20">20个金币</option>
+                                    <option value="-1">禁止下载</option>
                                 </select>
                             </div>
                         </div>
