+++ views/Home/default/User/edit.html
@@ -55,6 +55,7 @@
                                     <option value="8" {{if eq .Info.Price 8}}selected="selected"{{end}}>8个金币</option>
                                     <option value="10" {{if eq .Info.Price 10}}selected="selected"{{end}}>10个金币</option>
                                     <option value="20" {{if eq .Info.Price 20}}selected="selected"{{end}}>20个金币</option>
+                                    <option value="-1" {{if eq .Info.Price -1}}selected="selected"{{end}}>禁止下载</option>
                                 </select>
                             </div>
                         </div>
